combined interests and topics to avoid
