+++ src/pages/_pages_shared/Heading.sc.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 const Heading = styled.h1`
   width: 100%;
   font-size: 1.3em;
-  line-height: 150%;
+  line-height: 120%;
   text-align: center;
   font-weight: 700;
   margin: 0;
